refactor: split premium material armor renderer
